+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 20.5.11
 - some startup performance improvements were made (for [#1755](https://github.com/pbek/QOwnNotes/pull/1755), thank you @Waqar144)
+- todo list items now have a square as list-style (for [#1761](https://github.com/pbek/QOwnNotes/issues/1761))
 
 ## 20.5.10
 - fixed file links interpreted in code blocks (for [#1748](https://github.com/pbek/QOwnNotes/issues/1748), thank you @Waqar144)